ExploringLLMs/GoogleLLMs/googleOdet/requirements.txt
google-genai
python-dotenv
Pillow